+++ raster/src/main/scala/geotrellis/raster/resample/MinResample.scala
@@ -28,10 +28,14 @@ import scala.math
 class MinResample(tile: Tile, extent: Extent, targetCS: CellSize)
     extends AggregateResample(tile, extent, targetCS) {
 
-  private def calculateMin(indices: Seq[(Int, Int)]): Int =
-    indices.foldLeft(Int.MaxValue) { case (currentMin, coords) =>
-      math.min(currentMin, tile.get(coords._1, coords._2))
+  private def calculateMin(indices: Seq[(Int, Int)]): Int = {
+    val intMin = indices.foldLeft(Int.MaxValue) { case (currentMin, coords) =>
+      val v = tile.get(coords._1, coords._2)
+      // NODATA would *always* be min
+      if (isData(v)) math.min(currentMin, v) else currentMin
     }
+    if (intMin == Int.MaxValue) NODATA else intMin
+  }
 
   private def calculateMinDouble(indices: Seq[(Int, Int)]): Double = {
     val doubleMin = indices.foldLeft(Double.MaxValue) { case (currentMin, coords) =>